Add option to clean old mod files
The application can delete old mods before generating new ones. You can turn it on by setting `DeleteOldMods` to `true` in the `appsettings.Mine.json` file.
It will delete all the files in the `~mods` folder starting with value specified in `OutputPakName`. The default value is `ZonaBundle`. It can be useful if you turn on and off mods with additional files like assets.